refactor: split DatabaseManager into internal components
